Vishal | Code refactoring

+++ pojo/src/main/java/com/stanzaliving/boq_service/dto/PropertyBoqListingGroupByBoq.java
@@ -53,5 +53,5 @@ public class PropertyBoqListingGroupByBoq {
 
     private boolean canApprove;
 
-    private int commentCount;
+    private long commentCount;
 }
